+++ views/tags/_form.php
@@ -49,7 +49,7 @@ use wdmg\widgets\AliasInput;
         ])->label(Yii::t('app/modules/blog', 'Tag URL')); ?>
 
         <?php
-            if (isset(Yii::$app->redirects) && $model->url && ($model->status == $model::STATUS_PUBLISHED)) {
+            if (isset(Yii::$app->redirects) && $model->url) {
                 if ($url = Yii::$app->redirects->check($model->url, false)) {
                     echo Html::tag('div', Yii::t('app/modules/redirects', 'For this URL is active redirect to {url}', [
                         'url' => $url